Rake Automation + 1st Theme
The repository is now fully automated by a Rakefile which will treat
any folder starting with `css__` as a theme folder, and will:

* Convert its main SCSS source `styles.scss` in the `sass/` subfolder
  to `css/styles.css`.
* Convert all the shared PML sample/test docs from the `samples/` folder
  (in the repo root) to HTML in the theme root folder, linking to the
  theme's own stylesheet (`css/styles.css`).
* Convert to HTML any PML documents inside the theme folder.

The shared sample/test docs in the `samples/` folder are also converted
to HTML, in place, using the default stylesheet, so that they might be
used to compare custom stylesheets with the native PML CSS.

The `css__default-new/` contains the default PML 4.0.0 stylesheets,
adapted as a Sass/SCSS partials (the print version being currently
disabled). The goal is to create a "Sassifyed" version of the default
stylesheet, with some improvements.