Apply Modular Scale
Now the stylesheet uses modular scale, exposed via global CSS variables
(on :root) which end users can re-use to extend the stylesheet, or modify
their values to change the scale ratio and/or base font-size.

The ToC was improved, reducing font-sizes, and ensuring that a ToC
position at "top" has a nice frame to distinguish it from the main text.

Start documenting the work done so far, along with the pending taks,
within the README of the New Default Theme.